Feat : RiderLocationRepository 추상화
